Working on allowing for a simple 'io' import.
It seems to work linking the output objects via gcc I get a segfault when
trying to run the resulting executable.  Might be a data type issue passing the
LLVM string to the printf C function but might also be something wrong with the
linking process.

Also includes improvements to how strings are handled.